Inserção de autenticação e nova organização.

+++ database/conexao.js
@@ -16,9 +16,10 @@ async function login(usuario) {
     console.log("Verificando login");
     const conexaoAtiva = await conecta();
     const [resultado] = await conexaoAtiva.query("SELECT * FROM usuario WHERE login = ?;", [usuario.login]);
-    if (resultado.length == 0) { return "Usuário não cadastrado!" }
-    if (resultado[0].senha != usuario.senha) { return "Senha informada não confere!"}
-    return response = {status: "conectado", admin: resultado[0].admin == 1 ? true : false  };
+    // if (resultado.length == 0) { return 400 }
+    // if (resultado[0].senha != usuario.senha) { return "Senha informada não confere!"}    
+    // return response = {status: "conectado", admin: resultado[0].admin == 1 ? true : false  };
+    return resultado[0];
 }
 
 //usuarios
@@ -33,7 +34,8 @@ async function selecionaUsuario(id) {
     console.log(`Selecionado o usuário com id: ${id}`);
     const conexaoAtiva = await conecta();
     const [resultado] = await conexaoAtiva.query("SELECT * FROM usuario WHERE id=?;", [id]);
-    return resultado;
+    if (resultado.length == 0) {return "Não existe usuário com o ID informado!"}
+    return resultado[0];
 }
 
 async function insereUsuario(usuario) {
@@ -49,12 +51,16 @@ async function insereUsuario(usuario) {
 async function excluiUsuario(id) {
     console.log(`Apagando o usuário com id:${id}`);
     const conexaoAtiva = await conecta();
+    const [resultado] = await conexaoAtiva.query("SELECT * FROM usuario WHERE id=?;", [id]);
+    if (resultado.length == 0) {return "Não existe usuário com o ID informado!"}
     return await conexaoAtiva.query("DELETE FROM usuario WHERE id=?", [id]);
 }
 
 async function alteraUsuario(usuario) {
     console.log("Alterando usuário: " + usuario.nome);    
     const conexaoAtiva = await conecta();    
+    const [resultado] = await conexaoAtiva.query("SELECT * FROM usuario WHERE id=?;", [usuario.id]);
+    if (resultado.length == 0) {return "Não existe usuário com o ID informado!"}
     const sql = "UPDATE usuario SET nome = ?, senha = ?, admin = ? WHERE id = ?;";
     const parametros = [usuario.nome, usuario.senha, usuario.admin, usuario.id];
     return await conexaoAtiva.query(sql, parametros);
@@ -64,15 +70,18 @@ async function alteraUsuario(usuario) {
 async function listaTodosContatos(user_id) {
     console.log("Listando todos os contatos do usuário com id: " + user_id);
     const conexaoAtiva = await conecta();
-    const [resultado] = await conexaoAtiva.query("SELECT * FROM contato WHERE user_id=?;", [user_id]);
-    return resultado;
+    const [resultado] = await conexaoAtiva.query("SELECT * FROM usuario WHERE id=?;", [user_id]);
+    if (resultado.length == 0) {return "Não existe usuário com o ID informado!"}
+    const [resultado2] = await conexaoAtiva.query("SELECT * FROM contato WHERE user_id=?;", [user_id]);
+    return resultado2;
 }
 
 async function selecionaContato(id) {
     console.log(`Selecionado o contato com id: ${id}`);
     const conexaoAtiva = await conecta();
     const [resultado] = await conexaoAtiva.query("SELECT * FROM contato WHERE id=?;", [id]);
-    return resultado;
+    if (resultado.length == 0) {return "Não existe contato com o ID informado!"}
+    return resultado[0];
 }
 
 async function insereContato(contato) {
@@ -86,12 +95,16 @@ async function insereContato(contato) {
 async function excluiContato(id) {
     console.log(`Apagando o contato com id:${id}`);
     const conexaoAtiva = await conecta();
+    const [resultado] = await conexaoAtiva.query("SELECT * FROM contato WHERE id=?;", [id]);
+    if (resultado.length == 0) {return "Não existe contato com o ID informado!"}
     return await conexaoAtiva.query("DELETE FROM contato WHERE id=?", [id]);
 }
 
 async function alteraContato(contato) {
     console.log("Alterando usuário: " + contato.nome);
     const conexaoAtiva = await conecta();
+    const [resultado] = await conexaoAtiva.query("SELECT * FROM contato WHERE id=?;", [contato.id]);
+    if (resultado.length == 0) {return "Não existe contato com o ID informado!"}
     const sql = "UPDATE contato SET nome = ?, email = ?, telefone = ?, endereco = ?, user_id = ? WHERE id = ?;";
     const parametros = [contato.nome, contato.email, contato.telefone, contato.endereco, contato.user_id, contato.id];
     return await conexaoAtiva.query(sql, parametros);
@@ -101,15 +114,18 @@ async function alteraContato(contato) {
 async function listaTodosCompromissos(user_id) {
     console.log("Listando todos os compromissos do usuário com id: " + user_id);
     const conexaoAtiva = await conecta();
-    const [resultado] = await conexaoAtiva.query("SELECT * FROM compromisso WHERE user_id=?;", [user_id]);
-    return resultado;
+    const [resultado] = await conexaoAtiva.query("SELECT * FROM usuario WHERE id=?;", [user_id]);
+    if (resultado.length == 0) { return "Não existe usuário com o ID informado!" }    
+    const [resultado2] = await conexaoAtiva.query("SELECT * FROM compromisso WHERE user_id=?;", [user_id]);
+    return resultado2;
 }
 
 async function selecionaCompromisso(id) {
     console.log(`Selecionado o compromisso com id: ${id}`);
     const conexaoAtiva = await conecta();
     const [resultado] = await conexaoAtiva.query("SELECT * FROM compromisso WHERE id=?;", [id]);
-    return resultado;
+    if (resultado.length == 0) { return "Não existe compromisso com o ID informado!" } 
+    return resultado[0];
 }
 
 async function insereCompromisso(compromisso) {
@@ -123,12 +139,16 @@ async function insereCompromisso(compromisso) {
 async function excluiCompromisso(id) {
     console.log(`Apagando o compromisso com id:${id}`);
     const conexaoAtiva = await conecta();
+    const [resultado] = await conexaoAtiva.query("SELECT * FROM compromisso WHERE id=?;", [id]);
+    if (resultado.length == 0) { return "Não existe compromisso com o ID informado!" } 
     return await conexaoAtiva.query("DELETE FROM compromisso WHERE id=?", [id]);
 }
 
 async function alteraCompromisso(compromisso) {
     console.log("Alterando compromisso: " +  compromisso.id);
     const conexaoAtiva = await conecta();
+    const [resultado] = await conexaoAtiva.query("SELECT * FROM compromisso WHERE id=?;", [compromisso.id]);
+    if (resultado.length == 0) { return "Não existe compromisso com o ID informado!" } 
     const sql = "UPDATE compromisso SET data = ?, obs = ?, participantes = ?, endereco = ?,  status = ?, user_id = ? WHERE id = ?;";
     const parametros = [compromisso.data, compromisso.obs, compromisso.participantes, compromisso.endereco, compromisso.status, compromisso.user_id, compromisso.id];
     return await conexaoAtiva.query(sql, parametros);
